feat(frontend): make PaymentSuccessAnimation RSC-ready, localized, and lighter
Splits PaymentSuccessAnimation into a server-safe wrapper (PaymentSuccessAnimation.tsx)
and the interactive implementation (PaymentSuccessAnimationClient.tsx), so the
"use client" boundary sits at the animation itself rather than forcing every
importer into client-only code (#1175).

Adds PaymentSuccessSkeleton as the next/dynamic loading fallback, replacing the
bare Suspense fallback={null} with a real placeholder while the client bundle
streams in (#1176).

Adds real payment/common translations to en/es/pt messages, replacing the
hardcoded English fallback strings the component previously fell back to (#1177).

Migrates the animation from framer-motion to the lighter motion/react package and
removes the unused confetti and react-confetti dependencies, which were never
imported anywhere in the codebase (#1178).

Also fixes a real bug found while touching this component: usePrefersReducedMotion
computed its initial state via useEffect, so the confetti-trigger effect (which
runs in the same mount commit) always saw the stale default before the OS
preference loaded, meaning reduced-motion users still got a confetti burst on
first render. Switched to a lazy useState initializer that reads matchMedia
synchronously.
